Adds some controlled components
Side note, multiselect is not intuitive. It breaks react's "everything
is just a value" idea to hae to reference the html nodes directly.
Stumbling block was that event.target.value for multiselect just tells
you what the user recently clicked, not what the user's entire selection
is.  I can see that being useful, but it seems secondary to finding all
the values selected by the user in my opinion.