Fun with elevating state
Note that I ran into some issues with this example. Unintuitively for me
the constructor of a react component appears to only be called on the
first render, not on subsequent rerenders of the component based upon
passed in state. This led to surprising behavior since I put the
sentence reversing logic in the constructor thinking it'd be called
whenever I modified one fo the inputs I was watching. What happened in
effect was I had frozen the input value to be what was defined on the
components initialization.